Add pushing data to global var observer
